Add files via upload
A voice recorder is found in every smartphone and computer today. It is an application that is used to record sound and save it in a specific file format, which can be listened to and transferred to another device.

In the above code, I have defined a Python function to record and save your recorded files. It takes two parameters:

The first parameter is seconds, where you will enter the number of seconds you want to record your voice.
The second parameter is the file, where you will input the name by which you want your recorded file to be saved. For example, “voice.wav”.
After running the above code, it will show you a message that the recording has started, and after the number of seconds that you have given as input, it will show you that the recording is complete. Once completed, it will automatically be saved to the same directory where your Python file is located.