Add files via upload
The number guessing game is a popular game among programmers. In the number-guessing game, the program selects a random number between two numbers, and the user guesses the correct number.
In this game, the program selects a random number between two numbers, and the user guesses the correct number.
To create a guessing game, write a program to select a random number between 1 and 10. To give hints to the user, we can use conditional statements to tell the user if the guessed number is smaller, greater than or equal to the randomly selected number.