Add files via upload
The datasets that you find on the internet from various data sources are either created by companies and organizations or are collected from websites. You must have scraped data from web pages by using the Python libraries but may have stuck while preparing the scraped data to create a dataset.

For this program, I will be using the BeautifulSoup library in Python. Here I am going to search for a random term on Google and then I will collect the data from the very first page that Google shows me.

So, I searched for “comparison of programming languages” on Google and got this article as the first result. Let’s see how we can scrape data from this web page to create a dataset. Below is how we can use the BeautifulSoup library in Python for the task of web scraping to create a dataset.